requirements.txt
altgraph==0.16.1
bgtunnel==0.4.1
Click==7.0
et-xmlfile==1.0.1
jdcal==1.4.1
mailer==0.8.1
mysqlclient==1.3.13
openpyxl==2.6.4
pkg-resources==0.0.0
PyInstaller==3.6
python-dotenv==0.10.1
titlecase==0.12.0
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
